+++ gemcutter/Resource/Texture.h
@@ -102,7 +102,7 @@ namespace gem
 		void Remove(unsigned unit);
 		void Clear();
 
-		unsigned Size() const;
+		size_t Size() const;
 
 		// Returns the buffer bound at the specified unit.
 		Texture& operator[](unsigned unit);
